+++ cypress/support/step-definitions/decimal-steps.js
@@ -1,33 +1,5 @@
 import { commonDataElementInputPreviewNoIframe } from "../../locators";
 
-Then("Decimal component is disabled", () => {
-  commonDataElementInputPreviewNoIframe().should("have.attr", "disabled");
-  commonDataElementInputPreviewNoIframe()
-    .parent()
-    .should("have.attr", "disabled");
-});
-
-Then("Decimal component is enabled", () => {
-  commonDataElementInputPreviewNoIframe().should("not.have.attr", "disabled");
-  commonDataElementInputPreviewNoIframe()
-    .parent()
-    .should("not.have.attr", "disabled");
-});
-
-Then("Decimal component is readOnly", () => {
-  commonDataElementInputPreviewNoIframe().should("have.attr", "readonly");
-  commonDataElementInputPreviewNoIframe()
-    .parent()
-    .should("have.attr", "readonly");
-});
-
-Then("Decimal component is not readOnly", () => {
-  commonDataElementInputPreviewNoIframe().should("not.have.attr", "readonly");
-  commonDataElementInputPreviewNoIframe()
-    .parent()
-    .should("not.have.attr", "readonly");
-});
-
 When("I set Decimal input to {word}", (labelInput) => {
   commonDataElementInputPreviewNoIframe().clear().type(labelInput);
 });