+++ src/__experimental__/components/numeral-date/numeral-date.component.js
@@ -281,7 +281,7 @@ NumeralDate.propTypes = {
   /** Label */
   label: PropTypes.string,
   /** Text applied to label help tooltip */
-  labelHelp: PropTypes.string,
+  labelHelp: PropTypes.node,
   /** When true, label is placed in line with an input */
   labelInline: PropTypes.bool,
   /** Label alignment. Works only when labelInline is true */