+++ src/components/select/filterable-select/filterable-select.component.js
@@ -379,6 +379,7 @@ const FilterableSelect = React.forwardRef(
       }
 
       isInputFocused.current = false;
+      setOpen(false);
 
       if (onBlur) {
         onBlur(event);
@@ -388,7 +389,10 @@ const FilterableSelect = React.forwardRef(
     function handleTextboxMouseDown(event) {
       isMouseDownReported.current = true;
 
-      if (event.target.attributes["data-element"].value === "input") {
+      if (
+        event.target.attributes["data-element"] &&
+        event.target.attributes["data-element"].value === "input"
+      ) {
         isMouseDownOnInput.current = true;
       }
     }